Move examples over to new mesa selector
